IndexedDB/tests/submissions/Opera/idbdatabase_transaction-invalid.htm

<!DOCTYPE html>
<title>IDBDatabase.transaction() - opening a transaction defaults to a read-only mode </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../support.js"></script>

<script>

var db,
  t = async_test(),
  open_rq = createdb(t);

open_rq.onupgradeneeded = function(e) {
    db = e.target.result;
    db.createObjectStore('readonly');
};
open_rq.onsuccess = function(e) {
    db = e.target.result;

    // XXX: Expect what exception?
    assert_throws({}, function() { db.transaction('readonly', 'yeah'); });

    t.done();
};

</script>

<div id="log"></div>
